% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_data.R
\docType{data}
\name{canada}
\alias{canada}
\title{canada}
\format{A data frame with 247 rows and 5 variables:
\describe{
  \item{h1}{Name of first-level (region)}
  \item{h2}{Leaf names of second-level (elementary_aggregate)}
  \item{h3}{Leaf names of third-level values (intermediate_aggregate)}
  \item{color}{colors in which the plot-regions will be filled}
  \item{weight}{CPI in percent of the overall total}
}}
\usage{
data(canada)
}
\description{
An example data.frame using Canadian Consumer Price Index (CPI) to demonstrate the
voronoiTree package
}
\examples{
data("canada")
head(canada)
}
\keyword{datasets}
