% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_data.R
\docType{data}
\name{ExampleGDP}
\alias{ExampleGDP}
\title{ExampleGDP}
\format{A data frame with 42 rows and 6 variables:
\describe{
  \item{h1}{Name of first-level (redundant)}
  \item{h2}{Leaf names of second-level (continents)}
  \item{h3}{Leaf names of third-level values (countries)}
  \item{color}{colors in which the plot-regions will be filled}
  \item{weight}{GDP values in percent of the overall total}
  \item{codes}{short labels used for overlays in plotting}
}}
\usage{
data(ExampleGDP)
}
\description{
An example data.frame using GDP data to demonstrate the
voronoiTree package
}
\examples{
data("ExampleGDP")
head(ExampleGDP)
}
\keyword{datasets}
