% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi_testdata_example.R
\name{voronoi_testdata_example}
\alias{voronoi_testdata_example}
\title{voronoi_testdata_example}
\usage{
voronoi_testdata_example()
}
\value{
returns a json-string as in the example
from https://bl.ocks.org/Kcnarf/fa95aa7b076f537c00aed614c29bb568
}
\description{
voronoi_testdata_example
}
\examples{
voronoi_testdata_example()
}
