% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_app.R
\name{vt_app}
\alias{vt_app}
\title{vt_app}
\usage{
vt_app(maxRequestSize = 50, ...)
}
\arguments{
\item{maxRequestSize}{(numeric) number defining the maximum allowed filesize (in megabytes)
for uploaded files, defaults to 50MB}

\item{...}{arguments (e.g \code{host}) that are passed through
\code{\link[voronoiTreemap]{vt_app}} when starting the shiny application}
}
\value{
starts the interactive graphical user interface which may be used to perform the
anonymisation process.
}
\description{
starts the graphical user interface developed with \emph{shiny}.
}
\examples{
\dontrun{
vt_app()
}
}
