% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_nodes.R
\name{vt_create_node}
\alias{vt_create_node}
\title{vt_create_node}
\usage{
vt_create_node(total_lab = "Total")
}
\arguments{
\item{total_lab}{name of the total level}
}
\value{
a Node
}
\description{
vt_create_node
}
\examples{
vt_create_node("Total")
}
