% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_d3.R
\name{vt_d3}
\alias{vt_d3}
\title{Voronoi Treemap in an htmlwidget}
\usage{
vt_d3(data, elementId = NULL, width = NULL, height = NULL, seed = NULL,
  title = NULL, legend = FALSE, legend_title = NULL, footer = NULL)
}
\arguments{
\item{data}{a correct json data object}

\item{elementId}{optional a custom elementId to be returned}

\item{width}{width of the widget}

\item{height}{height of the widget}

\item{seed}{if defined, the plot is fixed}

\item{title}{NULL or a string for the title}

\item{legend}{TRUE/FALSE if a legend should be printed}

\item{legend_title}{NULL or a string for the title of the legend}

\item{footer}{NULL or a string for the footer text}
}
\description{
Function to generate an htmlwidget with an htmlwidget
}
\note{
Remake of HowMuch.net's article 'The Global Economy by GDP' by _Kcnarf bl.ocks.org/Kcnarf/fa95aa7b076f537c00aed614c29bb568
}
\examples{
vt_d3(vt_export_json(vt_testdata()))
data(ExampleGDP)
gdp_json <- vt_export_json(vt_input_from_df(ExampleGDP))
vt_d3(gdp_json)
}
