% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_export_json.R
\name{vt_export_json}
\alias{vt_export_json}
\title{vt_export_json}
\usage{
vt_export_json(node, file)
}
\arguments{
\item{node}{a Node object generated by \code{\link{vt_create_node}} or returned
from  \code{\link{vt_add_nodes}}}

\item{file}{path where the json should be written to}
}
\description{
exports a node to suitable json required by voronoi javascript function
}
\examples{
n <- vt_create_node("Total")
n <- vt_add_nodes(n, refnode="Total",node_names=c("Asia","Europe"),
  colors=c("red","blue"))
n <- vt_add_nodes(n, refnode="Asia",node_names=c("China","Thailand"),
  weights=c(0.5, 0.8), codes=c("CN","TH"))
n <- vt_add_nodes(n, refnode="Europe",node_names=c("Netherlands","Austria"),
  weights=c(0.9, 1.1), codes=c("NL","AT"))
vt_export_json(node=n, file=tempfile())
}
