% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_input_from_df.R
\name{vt_input_from_df}
\alias{vt_input_from_df}
\title{vt_input_from_df}
\usage{
vt_input_from_df(inp, scaleToPerc = FALSE)
}
\arguments{
\item{inp}{a data.frame with specific format}

\item{scaleToPerc}{(logical) scale to percent}
}
\value{
a Node that can be written to json using \code{\link{vt_export_json}}
}
\description{
create a tree-structure from a data.frame
}
\examples{
## non yet
}
