% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi_testdata_example.R
\name{vt_testdata}
\alias{vt_testdata}
\title{vt_testdata}
\usage{
vt_testdata()
}
\value{
returns a json-string as in the example
from https://bl.ocks.org/Kcnarf/fa95aa7b076f537c00aed614c29bb568
}
\description{
vt_testdata
}
\examples{
vt_testdata()
}
