% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_export_json.R
\name{vt_export_json}
\alias{vt_export_json}
\title{vt_export_json}
\usage{
vt_export_json(node, file = NULL)
}
\arguments{
\item{node}{a Node object generated by \code{\link{vt_create_node}} or returned
from  \code{\link{vt_add_nodes}}}

\item{file}{path where the json should be written to, if \code{NULL}, the json is returned as
a character}
}
\value{
NULL or a character vector
}
\description{
exports a node to suitable json required by voronoi javascript function
}
\examples{
n <- vt_testdata()
vt_export_json(n)
vt_export_json(n, file=tempfile())
}
\seealso{
vt_create_node vt_add_nodes
}
